data_collection/backend/requirements.txt
# Phase 1 (existing)
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# Phase 2 (new - API)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Phase 3 (Supabase auth + database)
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
